+++ src/random/index.ts
@@ -1,19 +1,6 @@
 import { bigIntFromBytes } from "@oslojs/binary";
 
-export function random(): number {
-	const buffer = new ArrayBuffer(8);
-	const bytes = crypto.getRandomValues(new Uint8Array(buffer));
-
-	// sets the exponent value (11 bits) to 01111111111 (1023)
-	// since the bias is 1023 (2 * (11 - 1) - 1), 1023 - 1023 = 0
-	// 2^0 * (1 + [52 bit number between 0-1]) = number between 1-2
-	bytes[0] = 63;
-	bytes[1] = bytes[1]! | 240;
-
-	return new DataView(buffer).getFloat64(0) - 1;
-}
-
-export function generateRandomInteger(max: bigint): bigint {
+export function generateRandomInteger(random: RandomReader, max: bigint): bigint {
 	if (max < 2) {
 		throw new Error("Argument 'max' must be a positive integer larger than 1");
 	}
@@ -21,7 +8,13 @@ export function generateRandomInteger(max: bigint): bigint {
 	const shift = inclusiveMaxBitLength % 8;
 	const bytes = new Uint8Array(Math.ceil(inclusiveMaxBitLength / 8));
 
-	crypto.getRandomValues(bytes);
+	try {
+		random.read(bytes);
+	} catch (e) {
+		throw new Error("Failed to retrieve random bytes", {
+			cause: e
+		});
+	}
 
 	// This zeroes bits that can be ignored to increase the chance `result` < `max`.
 	// For example, if `max` can be represented with 10 bits, the leading 6 bits of the random 16 bits (2 bytes) can be ignored.
@@ -30,7 +23,13 @@ export function generateRandomInteger(max: bigint): bigint {
 	}
 	let result = bigIntFromBytes(bytes);
 	while (result >= max) {
-		crypto.getRandomValues(bytes);
+		try {
+			random.read(bytes);
+		} catch (e) {
+			throw new Error("Failed to retrieve random bytes", {
+				cause: e
+			});
+		}
 		if (shift !== 0) {
 			bytes[0] &= (1 << shift) - 1;
 		}
@@ -39,36 +38,25 @@ export function generateRandomInteger(max: bigint): bigint {
 	return result;
 }
 
-export function generateRandomIntegerNumber(max: number): number {
+export function generateRandomIntegerNumber(random: RandomReader, max: number): number {
 	if (max < 2 || max > Number.MAX_SAFE_INTEGER) {
 		throw new Error("Argument 'max' must be a positive integer larger than 1");
 	}
-	return Number(generateRandomInteger(BigInt(max)));
+	return Number(generateRandomInteger(random, BigInt(max)));
 }
 
-export function generateRandomString(length: number, alphabet: string): string {
+export function generateRandomString(
+	random: RandomReader,
+	alphabet: string,
+	length: number
+): string {
 	let result = "";
 	for (let i = 0; i < length; i++) {
-		result += alphabet[generateRandomIntegerNumber(alphabet.length)];
+		result += alphabet[generateRandomIntegerNumber(random, alphabet.length)];
 	}
 	return result;
 }
 
-export type AlphabetPattern = "a-z" | "A-Z" | "0-9" | "-" | "_";
-
-export function alphabet(...patterns: AlphabetPattern[]): string {
-	const patternSet = new Set<AlphabetPattern>(patterns);
-	let result = "";
-	for (const pattern of patternSet) {
-		if (pattern === "a-z") {
-			result += "abcdefghijklmnopqrstuvwxyz";
-		} else if (pattern === "A-Z") {
-			result += "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
-		} else if (pattern === "0-9") {
-			result += "0123456789";
-		} else {
-			result += pattern;
-		}
-	}
-	return result;
+export interface RandomReader {
+	read(bytes: Uint8Array): void;
 }